Combine CspPolicy into CspReport
CspReport was its only caller, and permits? carries a precondition only
CspReport satisfies: quoted sources name no host, so a same origin uri
'self' allows still answers false. Documented on the method.
